% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_trait_data}
\alias{example_trait_data}
\title{Example trait data}
\format{Data frame with 3 columns and 393 rows}
\source{
NM398 paper
}
\usage{
example_trait_data
}
\description{
Example trait data
}
\keyword{datasets}

