% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tree}
\alias{example_tree}
\title{An example tree}
\format{Phylogenetic tree with 379 tips and 378 internal nodes.}
\source{
NM398 paper
}
\usage{
example_tree
}
\description{
An example tree
}
\keyword{datasets}

