% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_test.R
\name{sm_test}
\alias{sm_test}
\title{Run SM test}
\usage{
sm_test(trait, tree, rep = 999)
}
\arguments{
\item{trait}{trait data to test}

\item{tree}{phylogeny to test}

\item{rep}{number of randomisations to perform}
}
\description{
Run SM test
}

